Making World matrix make more sense - introduced Scale and Rotation as coordinates on Node, these will cause a recalculation of the World matrix if changed or its content is changed.
Camera is no longer a Node - no need for it to be a node, but maintaining Position property which was the only property used from node.

+++ Source/Balder.Core/Math/Coordinate.cs
@@ -9,7 +9,7 @@ namespace Balder.Core.Math
 			
 		}
 
-		public Coordinate(float x, float y, float z)
+		public Coordinate(double x, double y, double z)
 		{
 			X = x;
 			Y = y;
